refactor(next): migrate branch next and prev
