+++ src/commands/original-commands/next-or-prev/index.ts
@@ -1,27 +1,13 @@
 #!/usr/bin/env node
 import chalk from "chalk";
 import { execSync } from "child_process";
-import yargs from "yargs";
-import AbstractCommand from "../../../lib/abstract_command";
 import { logErrorAndExit } from "../../../lib/utils";
 import Branch from "../../../wrapper-classes/branch";
 
-const args = {} as const;
-type argsT = yargs.Arguments<yargs.InferredOptionTypes<typeof args>>;
-export class NextCommand extends AbstractCommand<typeof args> {
-  static args = args;
-  public async _execute(argv: argsT): Promise<void> {
-    await nextOrPrev("next");
-  }
-}
-export class PrevCommand extends AbstractCommand<typeof args> {
-  static args = args;
-  public async _execute(argv: argsT): Promise<void> {
-    await nextOrPrev("prev");
-  }
-}
-
-async function nextOrPrev(nextOrPrev: "next" | "prev") {
+export async function nextOrPrev(
+  nextOrPrev: "next" | "prev",
+  silent: boolean
+): Promise<void> {
   const currentBranch = Branch.getCurrentBranch();
   if (currentBranch === null) {
     logErrorAndExit(`Not currently on branch, cannot find ${nextOrPrev}.`);
@@ -33,13 +19,17 @@ async function nextOrPrev(nextOrPrev: "next" | "prev") {
       : await currentBranch.getParentsFromGit();
 
   if (candidates.length === 0) {
-    console.log(chalk.yellow(`Found no ${nextOrPrev} branch`));
+    if (!silent) {
+      console.log(chalk.yellow(`Found no ${nextOrPrev} branch`));
+    }
     process.exit(1);
   }
   if (candidates.length > 1) {
-    console.log(chalk.yellow(`Found multiple possibilities:`));
-    for (const candidate of candidates) {
-      console.log(chalk.yellow(` - ${candidate.name}`));
+    if (!silent) {
+      console.log(chalk.yellow(`Found multiple possibilities:`));
+      for (const candidate of candidates) {
+        console.log(chalk.yellow(` - ${candidate.name}`));
+      }
     }
     process.exit(1);
   }
